+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Implemented writable book and written book, a number of related classes were added. Check the commit history for details.
 - (API) Added class `CustomItemComponentDataGenerator` which adds basic support for custom item.
 - (API) Added method `CreativeItemCategory.getNamedGroup()` to get a named group in a creative category by its name.
+- (API) Added class `CreativeItemGroups` which contains all the available creative item groups in vanilla.
 - Added support for in-plugin resource pack, which allows plugins to include their resource pack under `assets/resource_pack`. This
   is useful for plugins that have custom items or blocks that require a resource pack to be displayed correctly.
 - Players now can see each other's permission level in pause menu.